Break code out into some internal packages
This helps cleanup the main package, not fully finished yet
